Move wrapper back into sharded_impls
