+++ sharktank/sharktank/ops/sharded_impls.py
@@ -26,10 +26,103 @@ from ..types import (
 )
 from ..types.tensors import unbox_tensor
 from ._registry import AllOfType, AllOfExprsVariadic, IsOfType
-from .signatures import *
 from .shape import broadcast_dims, broadcast_dim, unbroadcast_dim
 from ..utils import longest_equal_range
 
+
+def import_and_wrap_signatures():
+    """
+    Import the signatures from .signatures and wrap then so the shards of their inputs tensors are on the same devices.
+    For unary ops, also pins the result if the input is pinned (e.g. for transpose).
+    """
+
+    def transfer_n_pin(f):
+        """
+        Create a wrapper for each operation.
+        """
+        from ..types import ShardedTensor
+        from typing import List, Tuple, Dict, Any
+
+        def unwrap_args(
+            items: Tuple | Dict[str, Any]
+        ) -> Tuple[List[int | List[int]], List[ShardedTensor]]:
+            t_i, t_vals = [], []
+            for i, arg in enumerate(items):
+                if isinstance(arg, ShardedTensor):
+                    t_i.append(i)
+                    t_vals.append(arg)
+                elif isinstance(arg, list) and all(
+                    isinstance(val, ShardedTensor) for val in arg
+                ):
+                    t_i.append([i] * len(arg))
+                    t_vals.extend(arg)
+            return t_i, t_vals
+
+        def rewrap_args(
+            items: Tuple | Dict, t_i: List[int | List[int]], t_vals: List[ShardedTensor]
+        ) -> Tuple[Tuple, Dict[str, Any]]:
+            i_lookup = (
+                list(range(len(items)))
+                if isinstance(items, tuple)
+                else list(items.keys())
+            )
+            new_items = list(items) if isinstance(items, tuple) else dict(items)
+
+            for i in t_i:
+                if isinstance(i, int):
+                    new_items[i_lookup[i]] = t_vals.pop(0)
+                else:  # List[int]
+                    _popped_vals = [t_vals.pop(0) for _ in range(len(i))]
+                    new_items[i_lookup[i[0]]] = items[i_lookup[i[0]]].__class__(
+                        _popped_vals
+                    )
+
+            if isinstance(new_items, list):
+                new_items = tuple(new_items)
+            return new_items
+
+        def func_wrapper(*args: Tuple, **kwargs: Dict[str, Any]):
+            t_i_args, t_vals_args = unwrap_args(args)
+            t_i_kwargs, t_vals_kwargs = unwrap_args(list(kwargs.values()))
+            t_vals = t_vals_args + t_vals_kwargs
+
+            t_vals = transfer_if_needed(*t_vals)
+
+            args = rewrap_args(args, t_i_args, t_vals[: len(t_vals_args)])
+            kwargs = rewrap_args(kwargs, t_i_kwargs, t_vals[len(t_vals_args) :])
+            res = f(*args, **kwargs)
+            if isinstance(res, ShardedTensor) and len(t_vals) > 0:
+                pinned = (
+                    res.pinned
+                    or (len(t_vals) == 1 and t_vals[0].pinned)
+                    or f.__name__ == "reshard_like"
+                )  # TODO: How to handle this case properly
+                res = res.clone(devices=t_vals[0].devices, pinned=pinned)
+            return res
+
+        if hasattr(f, "override"):  # Needed for ops like gelu_tanh_approximation
+            func_wrapper.override = f.override
+        return func_wrapper
+
+    do_not_wrap = {
+        "all_gather",
+        "all_reduce",
+        "replicate",
+        "index_copy_",
+        "index_put_",
+        "transfer_to_logical_device",
+    }
+
+    from . import signatures
+
+    for func_name in signatures.__all__:
+        func = getattr(signatures, func_name)
+        if func_name not in do_not_wrap:
+            func = transfer_n_pin(func)
+        globals()[func_name] = func
+
+
+import_and_wrap_signatures()
 
 def transfer_if_needed(*tensors: Tuple[ShardedTensor]) -> List[ShardedTensor]:
     """
